// web/src/app/core/api/generated/models/paged-response-of-problem-list-item-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProblemListItemResponse } from '../models/problem-list-item-response';
export interface PagedResponseOfProblemListItemResponse {
  items?: Array<ProblemListItemResponse>;
  pageNumber?: (number | string);
  pageSize?: (number | string);
  totalCount?: (number | string);
  totalPages?: (number | string);
}