// web/src/app/core/api/generated/models/problem-detail-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DifficultyLevel } from '../models/difficulty-level';
import { FunctionSignatureResponse } from '../models/function-signature-response';
import { ProblemExecutionMode } from '../models/problem-execution-mode';
import { ProblemSampleResponse } from '../models/problem-sample-response';
import { TagResponse } from '../models/tag-response';
export interface ProblemDetailResponse {
  constraintsMarkdown?: string;
  difficulty?: DifficultyLevel;
  executionMode?: ProblemExecutionMode;
  functionSignature?: (FunctionSignatureResponse | null);
  id?: (number | string);
  isSolved?: (boolean | null);
  judgeVersion?: (number | string);
  memoryLimitKb?: (number | string);
  publishedAt?: (string | null);
  samples?: Array<ProblemSampleResponse>;
  slug?: string;
  statementMarkdown?: string;
  tags?: Array<TagResponse>;
  timeLimitMs?: (number | string);
  title?: string;
}