// web/src/app/core/api/generated/models/problem-list-item-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DifficultyLevel } from '../models/difficulty-level';
import { TagResponse } from '../models/tag-response';
export interface ProblemListItemResponse {
  difficulty?: DifficultyLevel;
  id?: (number | string);
  isSolved?: (boolean | null);
  slug?: string;
  tags?: Array<TagResponse>;
  title?: string;
}